Move single datacenter indicator into datacenter selector
