+++ ui/packages/consul-ui/app/components/consul/datacenter/selector/index.hbs
@@ -2,52 +2,57 @@
   class="dcs"
   data-test-datacenter-menu
 >
-  <DisclosureMenu
-    aria-label="Datacenter"
-    @items={{sort-by 'Primary:desc' 'Local:desc' 'Name:asc' @dcs}}
-  as |disclosure|>
-    <disclosure.Action
-      {{on 'click' disclosure.toggle}}
-    >
-      {{@dc.Name}}
-    </disclosure.Action>
-    <disclosure.Menu as |panel|>
-      <DataSource
-        @src={{uri '/*/*/*/datacenters'}}
-        @onchange={{action (mut @dcs) value="data"}}
-      />
-        <panel.Menu as |menu|>
+  {{#if (gt @dcs.length 1)}}
+    <DisclosureMenu
+      aria-label="Datacenter"
+      @items={{sort-by 'Primary:desc' 'Local:desc' 'Name:asc' @dcs}}
+      data-test-datacenter-disclosure-menu
+    as |disclosure|>
+      <disclosure.Action
+        {{on 'click' disclosure.toggle}}
+      >
+        {{@dc.Name}}
+      </disclosure.Action>
+      <disclosure.Menu as |panel|>
+        <DataSource
+          @src={{uri '/*/*/*/datacenters'}}
+          @onchange={{action (mut @dcs) value="data"}}
+        />
           <p class="dcs-message">
             Datacenters shown in this dropdown are available through WAN Federation.
           </p>
-          {{#each menu.items as |item|}}
-            <menu.Item
-              aria-current={{if (eq @dc.Name item.Name) 'true'}}
-              class={{class-map
-                (array 'is-local' item.Local)
-                (array 'is-primary' item.Primary)
-              }}
-            >
-              <menu.Action
-                {{on 'click' disclosure.close}}
-                @href={{href-to '.' params=(hash
-                  dc=item.Name
-                  partition=undefined
-                  nspace=(if (gt @nspace.length 0) @nspace undefined)
-                )}}
+          <panel.Menu as |menu|>
+            {{#each menu.items as |item|}}
+              <menu.Item
+                aria-current={{if (eq @dc.Name item.Name) 'true'}}
+                class={{class-map
+                  (array 'is-local' item.Local)
+                  (array 'is-primary' item.Primary)
+                }}
               >
-                {{item.Name}}
-              {{#if item.Primary}}
-                <span>Primary</span>
-              {{/if}}
-              {{#if item.Local}}
-                <span>Local</span>
-              {{/if}}
-              </menu.Action>
-            </menu.Item>
-          {{/each}}
-        </panel.Menu>
-    </disclosure.Menu>
-  </DisclosureMenu>
+                <menu.Action
+                  {{on 'click' disclosure.close}}
+                  @href={{href-to '.' params=(hash
+                    dc=item.Name
+                    partition=undefined
+                    nspace=(if (gt @nspace.length 0) @nspace undefined)
+                  )}}
+                >
+                  {{item.Name}}
+                {{#if item.Primary}}
+                  <span>Primary</span>
+                {{/if}}
+                {{#if item.Local}}
+                  <span>Local</span>
+                {{/if}}
+                </menu.Action>
+              </menu.Item>
+            {{/each}}
+          </panel.Menu>
+      </disclosure.Menu>
+    </DisclosureMenu>
+  {{else}}
+    <li class="dc-name" data-test-datacenter-single>{{@dcs.firstObject.Name}}</li>
+  {{/if}}
 </li>
