+++ ui/packages/consul-ui/app/components/hashicorp-consul/index.hbs
@@ -83,23 +83,16 @@
     <a
       href={{href-to 'index' params=(hash peer=undefined)}}
     ><Consul::Logo /></a>
-    {{#if (eq @dcs.length 1)}}
-      <div class="dc-name" data-test-datacenter>
-        {{@dc.Name}}
-      </div>
-    {{/if}}
   </:home-nav>
 
   <:main-nav>
     <ul>
-      {{#if (gt @dcs.length 1)}}
         <Consul::Datacenter::Selector
           @dc={{@dc}}
           @partition={{@partition}}
           @nspace={{@nspace}}
           @dcs={{@dcs}}
         />
-      {{/if}}
       <Consul::Partition::Selector
         @dc={{@dc}}
         @partition={{@partition}}
