+++ ui/packages/consul-ui/app/components/hashicorp-consul/index.scss
@@ -52,7 +52,7 @@
   }
   .dc-name {
     color: rgb(var(--tone-gray-600));
-    margin-left: 18px;
+    padding: 3.25px 0px;
     font-weight: var(--typo-weight-semibold)
   }
 }
